+++ spec/models/claim/personal_data_scrubber_spec.rb
@@ -57,6 +57,7 @@ RSpec.describe Claim::PersonalDataScrubber, type: :model do
       expect(cleaned_claim.bank_sort_code).to be_nil
       expect(cleaned_claim.bank_account_number).to be_nil
       expect(cleaned_claim.building_society_roll_number).to be_nil
+      expect(cleaned_claim.banking_name).to be_nil
       expect(cleaned_claim.personal_data_removed_at).to eq(Time.zone.now)
     end
   end
@@ -83,6 +84,7 @@ RSpec.describe Claim::PersonalDataScrubber, type: :model do
       expect(cleaned_claim.bank_sort_code).to be_nil
       expect(cleaned_claim.bank_account_number).to be_nil
       expect(cleaned_claim.building_society_roll_number).to be_nil
+      expect(cleaned_claim.banking_name).to be_nil
       expect(cleaned_claim.personal_data_removed_at).to eq(Time.zone.now)
     end
   end